Starting to refactor for fun and for profit
